// ionic-bet/src/providers/api-client.generated/model/userDto.ts
/**
 * Template
 * Template Web API
 *
 * The version of the OpenAPI document: 1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { BetDto } from './betDto';


export interface UserDto { 
    id?: string;
    email?: string;
    password?: string;
    credit?: number;
    bets?: Array<BetDto>;
    access_token?: string;
}